refactor: carList 생성자 추가
