+++ src/main/java/racing/controller/CarController.java
@@ -20,7 +20,7 @@ public class CarController {
 
     public void run() {
         String carNames = carView.getCarNames();
-        CarList carList = new CarList(createCarList(carNames));
+        CarList carList = new CarList(carNames);
 
         int turn = carView.getTryCount();
 
@@ -30,17 +30,5 @@ public class CarController {
 
         carView.printGameLogs(gameLogs);
         carView.printWinner(carList);
-    }
-
-    private List<Car> createCarList(String carNames) {
-        String[] carNamesList = carNames.split(",");
-        List<Car> carList = new ArrayList<>();
-        RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
-        for (String carName : carNamesList) {
-            Car car = new Car(carName, randomNumberGenerator);
-            carList.add(car);
-        }
-
-        return carList;
     }
 }
