+++ src/test/java/racing/domain/CarListTest.java
@@ -6,7 +6,9 @@ import racing.generator.FalseNumberGenerator;
 import racing.generator.TrueNumberGenerator;
 
 import java.util.Arrays;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,7 +25,6 @@ public class CarListTest {
                         new Car("car2", falseNumberGenerator)
                 )
         );
-
     }
 
     @Test
